Added "failed" for easy exception control.
I have serious doubts about this method, but for now the interface will
use it. It requires that all methods set failed to false, and checks if
any of its used methods makes failed true. If its happens, it must stop
executing (for preventing other methods making it false again).